refactor database schema setup
